06.08

+++ src/main/java/project/web/data/domain/NativePage.java
@@ -31,7 +31,7 @@ public class NativePage {
     private LocalDate paDate; // 작성일자
 
     @Column(name = "pa_charge")
-    private Integer paCharge; // 수수료
+    private Double paCharge; // 수수료
 
     @Column(name = "pa_res")
     private Boolean paRes; // 예약 여부
